@@ -48,8 +48,8 @@
     "@testing-library/jest-dom": "^5.16.4",
     "@testing-library/react": "^12.1.4",
     "@testing-library/user-event": "^13.5.0",
-    "@types/react": "^17.0.40",
-    "@types/react-dom": "^17.0.13",
+    "@types/react": "^18.0.6",
+    "@types/react-dom": "^18.0.2",
     "@types/sanitize-html": "^2.6.2",
     "@types/semver": "^7.3.9",
     "@types/url-parse": "^1.4.8",
@@ -85,8 +85,8 @@
     "jest": "^27.5.1",
     "keytar": "^7.7.0",
     "pretty-bytes": "^6.0.0",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.0.0",
+    "react-dom": "^18.0.0",
     "react-router": "^6.3.0",
     "react-router-dom": "^6.3.0",
     "tslog": "^3.3.3",
